What's left now?
everything I did is in the change logs. I'm not sure if we need an input
class or not, that remains to be seen. Remaining are the HUD, maps
maker, bringing in all the assets, fire methods, enemy movement, and
probably other stuff that I can't think of off the top of my head.