Platforms!
Added the turret mechanic. I still need to kill the player if they fall
off the stage when falling.  Still only a single platform, if it goes
bye bye off the left side, not sure what happens. When you land on the
platform, the enemies just fly away and there's no way to trigger the
next encounter.  I NEED THOSE SPRITES. the platform sprite needs to be
cropped so we're not just floating on it, aaaaand the stamina drain is a
bit too slow.  That's all for now. Ta-ta!